+++ verification/curator-service/ui/src/components/new-case-form-fields/Demographics.tsx
@@ -2,7 +2,7 @@ import { Field, useFormikContext } from 'formik';
 import { Select, TextField } from 'formik-material-ui';
 
 import FormControl from '@material-ui/core/FormControl';
-import { FormikAutocomplete } from './FormikFields';
+import { FormikAutocomplete } from '../common-form-fields/FormikFields';
 import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
 import NewCaseFormValues from './NewCaseFormValues';
