+++ verification/curator-service/ui/src/components/new-case-form-fields/LocationForm.tsx
@@ -6,7 +6,7 @@ import { Location as Loc } from '../Case';
 import Location from './Location';
 import LocationOnIcon from '@material-ui/icons/LocationOn';
 import React from 'react';
-import { RequiredHelperText } from './FormikFields';
+import { RequiredHelperText } from '../common-form-fields/FormikFields';
 import Scroll from 'react-scroll';
 import { TextField } from 'formik-material-ui';
 import axios from 'axios';
