+++ verification/curator-service/ui/src/components/new-case-form-fields/Transmission.tsx
@@ -1,5 +1,5 @@
 import ChipInput from 'material-ui-chip-input';
-import { FormikAutocomplete } from './FormikFields';
+import { FormikAutocomplete } from '../common-form-fields/FormikFields';
 import NewCaseFormValues from './NewCaseFormValues';
 import React from 'react';
 import Scroll from 'react-scroll';
